// src/repo/types/user.type.ts
export interface CreateUserSettingInput {
  timezone?: string | undefined
  siteThemeId?: string | undefined
  siteLanguageId?: string | undefined
}
export interface UpdateUserSettingInput {
  timezone?: string | undefined
  defaultCurrencyTwoId?: string | undefined
  siteThemeId?: string | undefined
  siteLanguageId?: string | undefined
}